Split tests on different modules
