+++ tox.ini
@@ -27,6 +27,7 @@ set_env =
     upgrade: TESTFILE=test_upgrade.py
     external-access: TESTFILE=test_external_access.py
     dynamic-allocation: TESTFILE=test_dynamic_allocation.py
+    observability: TESTFILE=test_observability.py
     tls: TESTFILE=test_tls.py
 pass_env =
     PYTHONPATH
@@ -79,7 +80,7 @@ commands =
 
 
 
-[testenv:integration-{charm,trust,ha,upgrade,external-access,tls,dynamic-allocation}]
+[testenv:integration-{charm,trust,ha,upgrade,external-access,tls,dynamic-allocation,observability}]
 description = Run integration tests
 set_env =
     {[testenv]set_env}
